backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
sqlalchemy
asyncpg
geoalchemy2
shapely
herbie-data
xarray
numpy
pandas
pydantic
pydantic-settings
supabase
eccodes
cfgrib
matplotlib
rioxarray
rasterio
pillow
pyproj
cartopy
scipy
